+++ crates/matrix-sdk/src/authentication/oauth/account_management_url.rs
@@ -73,43 +73,82 @@ pub enum AccountManagementActionFull {
     CrossSigningReset,
 }
 
-/// Build the URL for accessing the account management capabilities, as defined
-/// in [MSC].
+/// Builder for the URL for accessing the account management capabilities, as
+/// defined in [MSC4191].
 ///
-/// # Arguments
+/// This type can be instantiated with [`OAuth::account_management_url()`] and
+/// [`OAuth::fetch_account_management_url()`].
 ///
-/// * `account_management_uri` - The URL to access the issuer's account
-///   management capabilities.
-///
-/// * `action` - The action that the user wishes to take.
-///
-/// # Returns
-///
-/// A URL to be opened in a web browser where the end-user will be able to
-/// access the account management capabilities of the issuer.
-///
-/// # Errors
-///
-/// Returns an error if serializing the action fails.
+/// [`AccountManagementUrlBuilder::build()`] and
+/// [`AccountManagementUrlBuilder::build_or_ignore_action()`] return a URL to be
+/// opened in a web browser where the end-user will be able to access the
+/// account management capabilities of the issuer.
 ///
 /// [MSC4191]: https://github.com/matrix-org/matrix-spec-proposals/pull/4191
-pub(crate) fn build_account_management_url(
-    mut account_management_uri: Url,
-    action: AccountManagementActionFull,
-) -> Result<Url, serde_html_form::ser::Error> {
-    let extra_query = serde_html_form::to_string(action)?;
+/// [`OAuth::account_management_url()`]: super::OAuth::account_management_url
+/// [`OAuth::fetch_account_management_url()`]: super::OAuth::fetch_account_management_url
+#[derive(Debug, Clone)]
+pub struct AccountManagementUrlBuilder {
+    account_management_uri: Url,
+    action: Option<AccountManagementActionFull>,
+}
 
-    // Add our parameters to the query, because the URL might already have one.
-    let mut full_query = account_management_uri.query().map(ToOwned::to_owned).unwrap_or_default();
-
-    if !full_query.is_empty() {
-        full_query.push('&');
+impl AccountManagementUrlBuilder {
+    /// Construct an [`AccountManagementUrlBuilder`] for the given URL.
+    pub(super) fn new(account_management_uri: Url) -> Self {
+        Self { account_management_uri, action: None }
     }
-    full_query.push_str(&extra_query);
 
-    account_management_uri.set_query(Some(&full_query));
+    /// Set the action that the user wishes to take.
+    pub fn action(mut self, action: AccountManagementActionFull) -> Self {
+        self.action = Some(action);
+        self
+    }
 
-    Ok(account_management_uri)
+    /// Get the serialized action, if any.
+    fn serialized_action(&self) -> Result<Option<String>, serde_html_form::ser::Error> {
+        self.action.as_ref().map(serde_html_form::to_string).transpose()
+    }
+
+    /// Build the URL with the given serialized action.
+    fn build_inner(self, serialized_action: Option<String>) -> Url {
+        let Some(serialized_action) = serialized_action else {
+            return self.account_management_uri;
+        };
+
+        // Add our parameters to the query, because the URL might already have one.
+        let mut account_management_uri = self.account_management_uri;
+        let mut full_query =
+            account_management_uri.query().map(ToOwned::to_owned).unwrap_or_default();
+
+        if !full_query.is_empty() {
+            full_query.push('&');
+        }
+        full_query.push_str(&serialized_action);
+
+        account_management_uri.set_query(Some(&full_query));
+
+        account_management_uri
+    }
+
+    /// Build the URL to present to the end user.
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if serializing the action fails.
+    pub fn build(self) -> Result<Url, serde_html_form::ser::Error> {
+        let serialized_action = self.serialized_action()?;
+        Ok(self.build_inner(serialized_action))
+    }
+
+    /// Build the URL to present to the end user and ignore errors while
+    /// serializing the action.
+    ///
+    /// Returns the URL without an action if the action fails to serialize.
+    pub fn build_or_ignore_action(self) -> Url {
+        let serialized_action = self.serialized_action().ok().flatten();
+        self.build_inner(serialized_action)
+    }
 }
 
 #[cfg(test)]
@@ -117,55 +156,56 @@ mod tests {
     use ruma::owned_device_id;
     use url::Url;
 
-    use super::{build_account_management_url, AccountManagementActionFull};
+    use super::{AccountManagementActionFull, AccountManagementUrlBuilder};
 
     #[test]
     fn test_build_account_management_url_actions() {
         let base_url = Url::parse("https://example.org").unwrap();
         let device_id = owned_device_id!("ABCDEFG");
 
-        let url =
-            build_account_management_url(base_url.clone(), AccountManagementActionFull::Profile)
-                .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url.clone()).build().unwrap();
+        assert_eq!(url, base_url);
+
+        let url = AccountManagementUrlBuilder::new(base_url.clone())
+            .action(AccountManagementActionFull::Profile)
+            .build()
+            .unwrap();
         assert_eq!(url.as_str(), "https://example.org/?action=org.matrix.profile");
 
-        let url = build_account_management_url(
-            base_url.clone(),
-            AccountManagementActionFull::SessionsList,
-        )
-        .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url.clone())
+            .action(AccountManagementActionFull::SessionsList)
+            .build()
+            .unwrap();
         assert_eq!(url.as_str(), "https://example.org/?action=org.matrix.sessions_list");
 
-        let url = build_account_management_url(
-            base_url.clone(),
-            AccountManagementActionFull::SessionView { device_id: device_id.clone() },
-        )
-        .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url.clone())
+            .action(AccountManagementActionFull::SessionView { device_id: device_id.clone() })
+            .build()
+            .unwrap();
         assert_eq!(
             url.as_str(),
             "https://example.org/?action=org.matrix.session_view&device_id=ABCDEFG"
         );
 
-        let url = build_account_management_url(
-            base_url.clone(),
-            AccountManagementActionFull::SessionEnd { device_id },
-        )
-        .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url.clone())
+            .action(AccountManagementActionFull::SessionEnd { device_id })
+            .build()
+            .unwrap();
         assert_eq!(
             url.as_str(),
             "https://example.org/?action=org.matrix.session_end&device_id=ABCDEFG"
         );
 
-        let url = build_account_management_url(
-            base_url.clone(),
-            AccountManagementActionFull::AccountDeactivate,
-        )
-        .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url.clone())
+            .action(AccountManagementActionFull::AccountDeactivate)
+            .build()
+            .unwrap();
         assert_eq!(url.as_str(), "https://example.org/?action=org.matrix.account_deactivate");
 
-        let url =
-            build_account_management_url(base_url, AccountManagementActionFull::CrossSigningReset)
-                .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url)
+            .action(AccountManagementActionFull::CrossSigningReset)
+            .build()
+            .unwrap();
         assert_eq!(url.as_str(), "https://example.org/?action=org.matrix.cross_signing_reset");
     }
 
@@ -173,19 +213,34 @@ mod tests {
     fn test_build_account_management_url_with_query() {
         let base_url = Url::parse("https://example.org/?sid=123456").unwrap();
 
-        let url =
-            build_account_management_url(base_url.clone(), AccountManagementActionFull::Profile)
-                .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url.clone())
+            .action(AccountManagementActionFull::Profile)
+            .build()
+            .unwrap();
         assert_eq!(url.as_str(), "https://example.org/?sid=123456&action=org.matrix.profile");
 
-        let url = build_account_management_url(
-            base_url,
-            AccountManagementActionFull::SessionView { device_id: owned_device_id!("ABCDEFG") },
-        )
-        .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url)
+            .action(AccountManagementActionFull::SessionView {
+                device_id: owned_device_id!("ABCDEFG"),
+            })
+            .build()
+            .unwrap();
         assert_eq!(
             url.as_str(),
             "https://example.org/?sid=123456&action=org.matrix.session_view&device_id=ABCDEFG"
         );
+    }
+
+    #[test]
+    fn test_build_account_management_url_or_ignore_action() {
+        let base_url = Url::parse("https://example.org").unwrap();
+
+        let url = AccountManagementUrlBuilder::new(base_url.clone()).build_or_ignore_action();
+        assert_eq!(url, base_url);
+
+        let url = AccountManagementUrlBuilder::new(base_url)
+            .action(AccountManagementActionFull::Profile)
+            .build_or_ignore_action();
+        assert_eq!(url.as_str(), "https://example.org/?action=org.matrix.profile");
     }
 }