+++ crates/matrix-sdk/CHANGELOG.md
@@ -221,6 +221,12 @@ simpler methods:
     call to `OAuth::finish_login()`.
   - `AuthorizationResponse`, `AuthorizationCode` and `AuthorizationError` are
     now private.
+- [**breaking**] - `OAuth::account_management_url()` and
+  `OAuth::fetch_account_management_url()` don't take an action anymore but
+  return an `AccountManagementUrlBuilder`. The final URL can be obtained with
+  `AccountManagementUrlBuilder::build()` or
+  `AccountManagementUrlBuilder::build_or_ignore_action()`.
+  ([#4831](https://github.com/matrix-org/matrix-rust-sdk/pull/4831))
 
 ## [0.10.0] - 2025-02-04